Break up _pyrepl tests
