+++ Lib/test/test_pyrepl/__main__.py
@@ -1,0 +1,4 @@
+import unittest
+from test.test_pyrepl import load_tests
+
+unittest.main()
